Add a base class to test with an Accessor
